Move PosixFileStream class to its own files Update UserDefault to work with FileStream Fix undefined openFileStream on platforms other than Win32
